Send out the job_place type information in Twitter message.
Refactored the code to generate the string for the enum to a seperate
module called HumanizeHelper and adjusted the places were it gets
accessed.
