+++ lib/elixir_jobs_web/views/email_view.ex
@@ -1,5 +1,5 @@
 defmodule ElixirJobsWeb.EmailView do
   use ElixirJobsWeb, :view
 
-  import ElixirJobsWeb.OfferView, only: [human_get_place: 2, human_get_type: 2]
+  import ElixirJobsWeb.HumanizeHelper, only: [human_get_place: 2, human_get_type: 2]
 end
